mel: only add wic.bmap, not wic.bz2
Plenty of machines are now adding wic to IMAGE_FSTYPES themselves. We don't
want or need to duplicate that, and IMAGE_FSTYPES is generally in the
machine's control anyway, so let them do it. Keep wic.bmap, as that's more of
a distro choice, and doesn't result in duplication.

JIRA: SB-8529

Signed-off-by: Christopher Larson <chris_larson@mentor.com>