dosfstools: Apply the patch for specific version of dosfstools
dosfstools has both GPLv3 recipe as well as GPLv2 recipe in poky.
If user sets INCOMPATIBLE_LICENSE = "GPLv3" the build was failing
because GPLv2 version of the recipe was picked and the patch failed
to apply for the same. This patch renames the file from dosftools_%.bbappend
to dosfstools_4.0.bbappend.

JIRA: SB-8683

Signed-off-by: Sujith Haridasan <Sujith_Haridasan@mentor.com>