Use an event handler to add wic.bmap to IMAGE_FSTYPES
If a wic image type is enabled, also enable wic.bmap. We use an event handler,
as we can't reference IMAGE_FSTYPES from inline python in the same variable,
and while we can reference WKS_FULL_PATH, then the value in the config
metadata would be different than in image recipes. This wouldn't normally be
a problem, but toaster is using the config metadata value.

JIRA: SB-8685

Signed-off-by: Christopher Larson <chris_larson@mentor.com>