openssh: work around clock_gettime kernel bug on Linux x32
This patch is from debian, see https://bugs.debian.org/849923.

On Linux x32, the clock_gettime VDSO currently falls back to the x86-64
syscall, so allow that as well as its x32 sibling.

Without this, ssh connections are immediately terminated when attempting to
use privsep sandbox (which is the default) on x32.

JIRA: SB-8593

Signed-off-by: Christopher Larson <chris_larson@mentor.com>