scripts: toaster: consolidate creation of custom.xml
Toaster is fed through the custom.xml file for local
layers and recipes info. This file was being created
through a mixture of shell scripting and python which
would make maintenance and code readability difficult
at the end of the day.
We now consolidate the creation of this custom.xml by
using a single python script and clean up the shell
script accordingly.

JIRA: SB-8614.

Signed-off-by: Awais Belal <awais_belal@mentor.com>