mel: add recursive dependency of do_image_complete on do_populate_lic
Only do_rootfs, do_image and do_image_wic have recrdeptask set to
do_populate_lic. When a package is specified using EXTRA_IMAGEDEPENDS,
it is not made part of rootFS (i.e it is not pulled in by any of do_image,
do_image_wic or do_rootfs) and is only required to build the final image
(i.e added to dependencies of do_build which doesn't has recrdeptask set
to do_populate_lic). So we need to pull in do_populate_lic for packages
specified using EXTRA_IMAGEDEPENDS. For this purpose add recursive
dependency of do_image_complete on do_populate_lic.

http://jira.alm.mentorg.com:8080/browse/SB-8951

Signed-off-by: Fahad Arslan <Fahad_Arslan@mentor.com>