mel-checkout: add script to check out mel content
Using the manifests available in the MEL installation where this script
resides, create the BUILD_DIR, populate it with bitbake, layers, etc, from
this MEL installation, and configure it via the usual setup-environment script
from meta-mentor. This script should not be run manually, but via the
`setup-environment` script in the root of the MEL install.

The manifests live in `mel_install_dir/manifests/`, with subdirs for each
piece of 'extra' or 'addon' functionality. If multiple manifests are in
a given directory, and the exact manifest is not specified, the script will
prompt to select one. The current manifest format is one checkout per line,
with each line holding, as tab-separated values, the checkout path, the commit
to check out, and the remotes to create in the git repository

JIRA: SB-9151

Signed-off-by: Christopher Larson <chris_larson@mentor.com>