release/setup-environment: add setup script which uses mel-checkout
This setup script runs mel-checkout to populate the layers+bitbake in a build
directory, then sources the meta-mentor setup script to configure the build
directory. This script should be sourced, just as the meta-mentor one is.

Arguments are provided to specify the build directory, base manifest, and
extra manifests, and all other arguments are passed along to the underlying
meta-mentor setup script as is.

JIRA: SB-9151

Signed-off-by: Christopher Larson <chris_larson@mentor.com>