archive-release: add new mel_layers component
A new function is created (and therefore a new available component for
RELEASE_ARTIFACTS) which ships git pack files for each of our components
(layers+bitbake), writes out manifest files for mel-checkout to use to set up
their git repositories, and ships a toplevel `setup-environment` script which
uses mel-checkout and then configures the build directory with the meta-mentor
setup scripts.

Remotes are added to the git repositories for any public layers, and extra
'upstream' remotes can be added for any forks.

JIRA: SB-9151

Signed-off-by: Christopher Larson <chris_larson@mentor.com>