archive-release: add mel_downloads artifact task
This function writes out downloads manifests mirroring the manifests written
by mel_layers, indexed by checksum, and creates individual tarballs with each
source file, named by checksum.

In the new release setup, we want the ability to install a BSP on top of the
baseline, rather than in its own isolated install, so we needed to resolve any
issues with file conflicts in the MEL install. For the layers, we chose to ship
pack files and set up checkouts using them when the user sets up a build
directory. This does the same thing for downloads, but without git. Downloads
are named by checksum in the install directory and referenced in downloads
manifests which are used to set up `DL_DIR` for the new project directory
alongside the layer checkouts.

This does now require use of a new script, `bitbake-layers dump-downloads`,
which is used to gather information about what downloads belong to which
layers, as this information is required to ensure they're associated with the
correct manifests. This will require changes for our automated builds. For
example:

```
$ echo 'BB_GENERATE_MIRROR_TARBALLS = "1"' >>conf/auto.conf
$ bitbake -c archive_release_downloads_all archive-release
$ bitbake-layers dump-downloads -m archive-release
$ bitbake archive-release
```

Unfortunately, it was not possible to gather the layer mapping information from
within a task in `archive_release_downloads.bbclass`, so this was the only
option to get it right.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>