bitbake-layers: add {gather,dump}-downloads commands
These commands determine what downloads belong to what layers, for a given
target or targets. They use variable tracking to determine which entries in
SRC_URI were added by which recipes or bbappends to ensure they're mapped to
the correct layer. Currently, if a given download is referenced by multiple
layers, the lowest priority layer is preferred, i.e. a download is associated
with oe-core rather than some random third party layer.

This is needed for the new mel downloads handling, to ensure the downloads are
tied to the correct downloads manifest, for e.g. updates and addons.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>