update_font_cache: disable for x86_64
This depends on use of bbpath-intercepts, which mel includes by default. There
is an issue running fc-cache under qemu user mode for x86_64 when doing
a build using the sourcery toolchain. Prior to the upgrade to qemu 2.10-rc2,
the postinst intercept just failed like so:

    qemu-x86_64: .../tmp/work/x86_64-linux/qemu-native/2.8.1.1-r0/qemu-2.8.1.1/translate-all.c:175: tb_lock: Assertion `!have_tb_lock' failed.
    qemu-x86_64: .../tmp/work/x86_64-linux/qemu-native/2.8.1.1-r0/qemu-2.8.1.1/translate-all.c:175: tb_lock: Assertion `!have_tb_lock' failed.

After the upgrade, qemu-x86_64 just hangs, preventing a successful build, so
disable this intercept and let it run on first boot, so it behaves like it did
prior to the qemu upgrade, until we can nail down the root cause.

This is specific to our toolchain, so I had considered putting it in
meta-sourcery, but the workaround depends on a bbclass in this layer, and
I want to avoid that sort of cross-layer dependency. Once the
bbpath-intercepts functionality goes upstream, I'll move this workaround to
meta-sourcery where it belongs.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>