archive-release: download and ship bmaptool
The bmap-tools project distributes a self-contained bmaptool binary which
includes the entirety of the bmap-tools python package, so we can download and
distribute that, or download and distribute a wrapper which does so, rather
than requiring the user install it or attempt to do so in the distro setup
scripts (i.e. setup-ubuntu).

This script is under the GPLv2, so redistribution shouldn't be a problem,
license wise.

JIRA: SB-10713

Signed-off-by: Christopher Larson <chris_larson@mentor.com>