udev/udev-extraconf: Fixed overlapping mounts
The automounter udev script attempts to mount the block devices under
/run/media/<partition-label> or under /run/media/<uuid>.
If two devices have same partition-label, they will be mounted on the
same location.
Also, the technique being used previously to evaluate UUID to be used
as the mount location is not right. It evaluates to the value of
whatever is on the second place in "blkid"'s output - which is not
always UUID. It can be SEC_TYPE, TYPE, LABEL or something else.

This causes two types of known issues:

1. MEL installation fails if a MEL image is already installed on a
media other than where attempted to install such as HDD/eMMC/SSD/USB
that is attached during installation.

2. Unmount fails on a device if attempted using the device node
e.g. "/dev/sda1" with the 'umount' command.

Changes in this commit fix such issues.

Associated issues: INTAMDDET-2258, INTAMDDET-2259

Signed-off-by: Arsalan-Awan <Arsalan_Awan@mentor.com>