consolekit: add dependency on var-volatile.mount for restart/stop
consolekit restart and stop loggers try to write to /var/volatile/log
before the target completes shutdown. Add an After= dependency on
var-volatile.mount unit so that they run before it is unmounted during
shutdown.

To achieve this, these services must be started at boot time rather than
just before shutdown, so that the order of start and stop is respected.
Instead of running the required functionality on ExecStart it must then
be run on ExecStop. In addition, since there is no dependency on
reboot/halt from ExecStop, a script finds out what shutdown path is run
and invokes the correct consolekit utility. Only one service is thus
needed, which handles all cases.

Signed-off-by: Benjamin Walsh <Benjamin_Walsh@mentor.com>