sdk_auto_relocate: run both aspects of enable/disable at the same time
There was a mismatch between when the logic which disables if lib is
empty (packaging time) and when the variables that disable it are obeyed
(parse time). Ensure that they run at the same time to resolve this
mismatch and ensure the reloc is fully disabled appropriately in all
cases where it should be.

JIRA: SB-10686

Signed-off-by: Christopher Larson <chris_larson@mentor.com>