toaster_configuration: fix layer paths
LAYER_PATHS is used for fetching each layer's details
as the name intends. The bblayers.conf now uses the
TOPDIR variable to depict the path for a layer whereas
earlier it was an absolute path. This change breaks
manipulation of LAYER_PATHS so we fix it by replacing
the variable with the actual path to conform with the
earlier mechanism.

Signed-off-by: Awais Belal <awais_belal@mentor.com>