toaster_configuration: define a bitbake version in custom fixture
The custom fixture now requires a bitbake version tied up
with every release definition and needs this bitbake version
to be defined in the fixture otherwise the tables are not
created properly and the fixture fails to load.
We now define an arbitrary bitbake version which is tied
to the MEL release definition to be loaded to fix the
table creation issue.

Signed-off-by: Awais Belal <awais_belal@mentor.com>