toaster_configuration: override TOASTER_DIR from cmdline
We'd like our build environment to be confined to the
BUILDDIR otherwise if not overridden here the toaster
binary expects and creates toaster artifacts at
`dirname BUILDDIR` which is one level up to where we'd
like it to be. The toaster binary does this solely
because it expects a directory structure identical to
poky which is not the case with MEL anymore.

Signed-off-by: Awais Belal <awais_belal@mentor.com>