mel: local.conf.append: allow gplv3 binutils for aarch64, for tools-profile
The gplv2 binutils does not support aarch64, so the gplv3 version is
built and installed with tools-profile on that architecture by default,
even in a non-gplv3 build. This is kept in local.conf next to the bits
which install binutils with tools-profile by default, to keep it both
user-visible and easy to disable.

JIRA: SB-11714

Signed-off-by: Christopher Larson <chris_larson@mentor.com>