setup-environment: handle ${TOPDIR} in BBLAYERS
In the case of the new git delivery mechanism, TOPDIR==MELDIR, so we end
up with `${TOPDIR}` references in BBLAYERS, but `configured_layers` in
setup-environment wasn't converting that to a reference to BUILDDIR, and
hence failed to find any layers for handling layer.conf.append or
post-setup-environment, resulting in much configuration missing from
local.conf. This broke builds of just about all of our BSPs.

JIRA: SB-11836

Signed-off-by: Christopher Larson <chris_larson@mentor.com>