setup-mel/setup-builddir: only setup a workspace
The setup-builddir script created a builddir itself if
one wasn't provided by the user and then it would call
setup-environment for it. This causes ambiguities where
a user would want to setup different builds using the
same workspace.
Now we drop the calls to setup-environment and do not
create a builddir from these scripts so the old mechanism
of calling setup-environment separately can be used as
earlier and avoid confusions. The setup-builddir script
is renamed to setup-workspace so the intent is clear.

Signed-off-by: Awais Belal <awais_belal@mentor.com>