local.conf.sample: drop x11-mini, clarify 'graphics' feature
Drop x11-mini from the features list as it is an old remnant
of cleanup. The feature is now known as x11-base but that is
explained in OE's sample files listed in the definition of
features here.
The graphics feature does not always default to wayland/weston
but rather that decision is taken looking at the MACHINE_FEATURES
which is what provided by the vendor support.

Signed-off-by: Awais Belal <awais_belal@mentor.com>