mel-checkout: only allow reusing workspace for same machine
If a user configures a workspace using one BSP and creates
build directories using the metadata inside, then tries to
reconfigure/update the same workspace mistakenly or intentionally
for a different BSP the metadata and already created build
directories would go out of sync in a way that would cause
unknown errors in most cases.
Now we only allow the user to reconfigure/update a workspace
for the same machine that was initially used to setup a particular
workspace. This mechanism would also allow us to provide the
machine name to setup-environment in the future without the
user having to specify it each time on the command line manually.

Signed-off-by: Awais Belal <awais_belal@mentor.com>