setup-workspace: rework argument handling
Originally, this script was a wrapper script, so used non-standard
argument handling to allow it to pass unrecognized arguments through to
the underlying script directly. Since it was reworked to only set up the
workspace, this is no longer the case, so the old argument handling is
not appropriate. Rework to use getopts and properly show errors/usage on
incorrect arguments.

JIRA: SB-11943

Signed-off-by: Christopher Larson <chris_larson@mentor.com>