setup-mel-builddir: set BBLAYERS using loose assignment
In cases such as toaster the setting for BBLAYERS is
done through cmdline params and then during the parse
phase bblayers.conf is picked up which overrides the
setting that was passed if we use a simple set (=)
here.
Fix this by using loose assignment so cmdline params
work as expected.

JIRA: SB-12082.

Signed-off-by: Awais Belal <awais_belal@mentor.com>