connman: Restrict connman from including it's DNS proxy
DNS proxy is needed by docker. And when docker comes
with the inclusion of virtualization-layer, dnsmasq also
comes. dnsmasq tries to bind to port 53, which is
already in use of connman, and this results in the
failure of dnsmasq service.
As with virtualization-layer dnsproxy is provided by
dnsmasq, and without virtualization-layer it is not
needed.
In any case dnsproxy is not required from connman,
so,run connman with --nodnsproxy to avoid this failure.

Jira: http://jira.alm.mentorg.com:8080/browse/SB-12022

Signed-off-by: abdulkarim17 <hafizabdul_karim@mentor.com>