toaster_configuration: set OE_ROOT for toaster template searching
The toaster binary expects user to set the OE_ROOT var in
case a directory structure different to yocto is used otherwise
the binary fails to load the template configuration file and
throws a bash error on the cmdline.
This difference of directory structure exists in MEL so we
need to export the variable explicitly.

JIRA: SB-12080.

Signed-off-by: Awais Belal <awais_belal@mentor.com>