local.conf: set encrypted-fs as a user feature
encrypted-fs was previously being used as an EXTRA_IMAGE_FEATURE.
But machine related kernel configurations cannot be made on the
basis of an image feature. Since in most of the imx bsps, we
encountered hardware crypto failure, kernel configurations needed
to be set on the basis of encrypted-fs. And that's why it is
to be set as USER FEATURE instead of IMAGE FEATURE.

This commit updates the checks that were previously being made
on the basis of IMAGE FEATURE and lets the user to add encrypted-fs
as USER_FEATURE just by uncommenting a line.

Jira Issue: http://jira.alm.mentorg.com:8080/browse/SB-12527

Signed-off-by: arshadaleem66 <arshad_aleem@mentor.com>