python3-docker-compose: add missing dependencies
docker-compose needs a number of python3 runtime dependencies. The
packages python3-misc and python3-modules provide those dependencies.
The bbappend is being applied to current version of python3-docker-
compose so that it is reviewed later on as well if these dependencies
are fixed by upstream in the later versions or not.

Signed-off-by: arshadaleem66 <arshad_aleem@mentor.com>