setup-environment: silence the second setup-environment, not the first
If there is any one-time setup needed, the user can be staring at an inactive
terminal for some time, so silence the second, not the first. The only potential
issue there is any new output from the post-setup-environment scripts would be
hidden. This may need to be revisited to rework how this is done, or do a
partial setup first, not a complete one.

Signed-off-by: Christopher Larson <kergoth@gmail.com>