setup-environment: remove new builddir cleanup
It's best to clean it up where it's created, in setup-mel-builddir, and
further, we need to minimize what's being done in the user's interactive
shell. Doing the rm in that context without `-f`, for example, can
result in it prompting the user every time if they use an rm alias that
passes `-i`.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>