mel.conf: disable sanity check from meta-virtualization
Fixes
WARNING: You have included the meta-virtualization layer, but 'virtualization' has not been enabled in your DISTRO_FEATURES. Some bbappend files may not take effect. See the meta-virtualization README for details on enabling virtualization support.

Our support for the layer is implemented a bit differently
as we only pick a handful of packages from it.

Signed-off-by: Awais Belal <awais_belal@mentor.com>