tripwire: deploy installation script to an appropriate location
The main recipe deploys the installation script to /etc which
isn't meant for executables and doesn't work well when a read-only
or no-exec strategy is in place for /etc which can be the case
in many systems. This renders the installation script invalid
to be used with such strategy.
We now deploy the script to bindir which is always an executable
location by any specification so it can be executed.

Signed-off-by: Awais Belal <awais_belal@mentor.com>