setup-mel-builddir: do not depend on mentor-bsp for every machine
We have a requirement where upstream Yocto BSPs should be able to
build with MEL out of the box. Sourcing setup-environment at this
point for an upstream machine adds the MEL common BSP layer to
bblayers automatically which is not correct as that particular
machine is not a MEL BSP. However, we do need this for qemu BSPs
as we use the same upstream machine name but create a layer for
our modifications.

Signed-off-by: Awais Belal <awais_belal@mentor.com>