local.conf.sample: conditionalize IMAGE_FEATURES_DEVELOPMENT
When building the development-image we enable various image
features by default irrespective of the BSP being built with
the mel distro. This causes various issues while building a
BSP which has not been fully migrated to use mel.
We now conditionally add some specific features to image
features for the development-image when the BSP advertises
that it is a mel supported BSP.

Signed-off-by: Awais Belal <awais_belal@mentor.com>