mel.conf: set SPLASH using our distro override
When psplash is set as the SPLASH, the splash screen
images can be manipulated with SPLASH_IMAGES. We use an
append to set SPLASH_IMAGES up for mel using an override.
Use the same override for setting SPLASH so we don't get
into a scenario where SPLASH and SPLASH_IMAGES differ in
the configuration.

Signed-off-by: Awais Belal <awais_belal@mentor.com>