setup-mel-builddir: conditionalize inclusion of machine layer on qemu
This was discussed under
https://github.com/MentorEmbedded/meta-mentor/pull/1379
as well but we missed a point where most of the BSP sub-layers
under meta-mentor-bsp are created with the name
mentor-bsp-<upstream_machine>. Now, if a user wants to build the
upstream machine under the MEL metadata they'd source the setup
script with the upstream machine name and our migrated machine
would automatically be picked up which is not something that we'd
want for such a usecase. However, we do need this for the qemu
machines as those are special.

Signed-off-by: Awais Belal <awais_belal@mentor.com>