mel_utils.py: add source and homepage info for dump-licenses
The dump-licenses command additionally processes SRC_URI and
HOMEPAGE for each of the packages. This is a requirement for
the Siemens SVM.
This change will have impact on the OSS jobs that generate
the diffs for OSS submissions as it utilizes the output of
dump-licenses already so modifications to that part should
be considered prior to merging this.

Signed-off-by: Awais Belal <awais_belal@mentor.com>