staging/openembedded-layer/nodejs: upgrade to 8.16.1
This upgrades the nodejs package to 8.16.1 from 8.9.4 to
resolve a number of CVEs that were reported against the
older version. The backporting effort for the CVE fixes
on the older version out weighs the work that we are
expecting here so we simply upgrade the package.

Fixed CVEs:
CVE-2019-9511
CVE-2019-9512
CVE-2019-9513
CVE-2019-9514
CVE-2019-9515
CVE-2019-9516
CVE-2019-9517
CVE-2019-9518

JIRA: https://jira.alm.mentorg.com/browse/SB-16893
Signed-off-by: Awais Belal <awais_belal@mentor.com>