mel_utils.py/dump-licenses: add flag to append output file
This adds --append flag to the dump-licenses function. Using this
we can run this function to gather license information of packages
and their dependencies from multiple images and recipe tasks like
populate_sdk etc, and have them added to a single file.

This also removes any duplicates in the output file added when this
function is run multiple times with append flag enabled, such as:

$ bitbake-layers dump-licenses development-image
$ bitbake-layers dump-licenses development-image -c populate_sdk -a

Signed-off-by: Arsalan H. Awan <Arsalan_Awan@mentor.com>