mel_utils/dump-licenses: support multi-config scenarios
Follows the same methodology that the previous commit used
for dump-downloads. It is identified that this philosophy
still does not cater fully to the multiconfig scenario
where different configs might use different sources (SRC_URI)
in which case the download and the license might be
different.

JIRA: https://jira.alm.mentorg.com/browse/SB-19283

Signed-off-by: Awais Belal <awais_belal@mentor.com>