mel: work around safe dir issue with METADATA_*
With recent versions of git, attempting to operate on a git repository owned by
one user as another user will fail. As our builds run under pseudo in some
contexts, this fails. This issue in tasks can be worked around with a git
wrapper script, and is worked around upstream, but that does not resolve the
issue with metadata_scm.bbclass, as it isn't run with the correct PATH. It's a
non-issue upstream as they already made this change to force immediate expansion
of these variables in the right context, so apply that here also.

JIRA: SB-19344

Signed-off-by: Christopher Larson <chris_larson@mentor.com>